pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1.44",
    "google-genai>=1.21.1",
    "requests>=2.32.4",
    "streamlit>=1.46.0",
]
